requirements.txt
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.4
Pillow==7.1.2
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.42
six==1.15.0
